more common resource organization and unit test organization
