+++ do_py/data_object/restriction.py
@@ -609,46 +609,6 @@ class ManagedRestrictions(object, with_metaclass(ABCMeta)):
         return self._restriction == other._restriction
 
 
-class ManagedList(ManagedRestrictions):
-    _restriction = Restriction([list, type(None)])
-
-    def __init__(self, obj_cls, nullable=False):
-        super(ManagedList, self).__init__()
-        self.obj_cls = obj_cls
-        self.nullable = nullable
-
-    def manage(self):
-        if self.data is not None:
-            items = []
-            for item in self.data:
-                items.append(item if type(item) == self.obj_cls else self.obj_cls(item))
-            self.data = items
-        else:
-            if not self.nullable:
-                raise RestrictionError.bad_data(self.data, self._restriction.allowed)
-
-
-class OrderedManagedList(ManagedList):
-
-    def __init__(self, obj_cls, nullable=False, key=None, reverse=False):
-        """
-        :param obj_cls: DataObject class reference to wrap each object in list.
-        :type nullable: bool
-        :type key: function
-        :type reverse: bool
-        """
-        self.key = key
-        self.reverse = reverse
-        super(OrderedManagedList, self).__init__(obj_cls, nullable=nullable)
-
-    def manage(self):
-        """
-        Sort the data list after ManagedList does its work.
-        """
-        super(OrderedManagedList, self).manage()
-        self.data = sorted(self.data, key=self.key, reverse=self.reverse)
-
-
 # NOTE: ESR should be refactored to common.py, but OOS for now due to circular imports with ESEncoder
 class ESR(object):
     """
