+++ tests/test_common/test_managed_datetime.py
@@ -1,30 +1,14 @@
 """
-:date_created: 2019-07-25
+
+:date_created: 2020-06-28
 """
 
 from datetime import datetime
 
 import pytest
-from do_py import DataObject
 
+from do_py.common.managed_datetime import MgdDatetime
 from do_py.exceptions import RestrictionError
-
-from do_py.data_object.common import MgdDatetime, R
-from do_py.data_object.restriction import ESR
-
-
-class TestR(object):
-    def test_r_constants(self):
-        for attr in [r for r in dir(R) if not r.startswith('__')]:
-            assert getattr(R, attr)
-
-    @pytest.mark.parametrize('args, kwargs', [
-        ([], {}),
-        ([int], {}),
-        ([int], {'default': 1}),
-        ])
-    def test_r_creation(self, args, kwargs):
-        assert R(*args, **kwargs)
 
 
 test_dt_instance_epoch = datetime.fromtimestamp(0).replace(microsecond=0)
@@ -100,50 +84,3 @@ class TestMgdDatetime(object):
         """
         instance = MgdDatetime.null_date()
         assert output == instance(input)
-
-
-class TestESR(object):
-    """
-    Tests the util ESR, the compilation of ElasticSearch's mapping declaration structure at restriction level.
-    """
-
-    def test_constants(self):
-        """
-        Tests the static mapping values housed in ESR.
-        """
-        for attr in [r for r in dir(ESR) if not r.startswith('__')]:
-            assert getattr(ESR, attr)
-
-    @pytest.fixture(params=[True, pytest.param(False, marks=pytest.mark.xfail(raises=RestrictionError))])
-    def data_object(self, request):
-        """
-        :rtype: DataObject
-        """
-        class SampleDO(DataObject):
-            _restrictions = {
-                'x': R.INT
-                }
-            es_restrictions = {
-                'x': ESR.INT
-                }
-
-        class BadSampleDO(DataObject):
-            _restrictions = {
-                'x': R.INT
-                }
-
-        return SampleDO({'x': 1}) if request.param else BadSampleDO
-
-    def test_object(self, data_object):
-        """
-        Tests the dynamic mapping value creation from a DataObject for nested object mappings.
-        :type data_object: DataObject
-        """
-        assert ESR.OBJECT(data_object)
-
-    def test_nested(self, data_object):
-        """
-        Tests the dynamic mapping value creation from a DataObject for nested (object list) mappings.
-        :type data_object: DataObject
-        """
-        assert ESR.NESTED(data_object)
