+++ tests/test_restrictions.py
@@ -10,8 +10,8 @@ from datetime import date, datetime
 import pytest
 
 from do_py import DataObject
-from do_py.data_object.common import R
-from do_py.data_object.restriction import AbstractRestriction, ManagedRestrictions, Restriction, SingletonRestriction, \
+from do_py.common import R
+from do_py.data_object.restriction import AbstractRestriction, ManagedRestrictions, SingletonRestriction, \
     _DataObjectRestriction, \
     _ListTypeRestriction, _ListValueRestriction, _MgdRestRestriction, _NullableDataObjectRestriction
 from do_py.exceptions import DataObjectError, RestrictionError
